pyproject.toml
[project]
name = "xmem-mcp"
version = "1.0.0"
description = "MCP server for long-term memory with XMem"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["httpx>=0.28.1", "mcp[cli]>=1.3.0", "python-dotenv>=1.0.0"]
